move i18n code outside app dir, add types
