+++ apps/ui/package.json
@@ -20,6 +20,7 @@
     "postcss": "8.4.23",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "tailwindcss": "3.3.2"
+    "tailwindcss": "3.3.2",
+    "zod": "^3.21.4"
   }
 }
