+++ apps/ui/src/app/[lang]/layout.tsx
@@ -16,10 +16,12 @@
  * IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-
-import { ValidLanguage, supportedLanguages } from '../i18n/settings';
-import '../globals.css';
 import { Montserrat } from 'next/font/google';
+
+import '../globals.css';
+
+import { ValidLanguage } from '@/i18n';
+import { supportedLanguages } from '@/i18n/settings';
 
 const inter = Montserrat({ subsets: ['latin'] });
