+++ apps/ui/src/i18n/hooks/getTranslation.ts
@@ -17,40 +17,24 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import {
-  ValidLanguage,
-  ValidNamespace,
-  defaultLanguage,
-  defaultNS,
-} from './settings';
+import { GetDictionary, GetTranslation } from '../types';
+import { defaultLanguage, defaultNamespace } from '../settings';
 
-type GetDictionary = {
-  [k in ValidLanguage]: (
-    namespace: ValidNamespace
-  ) => Promise<{ [k: string]: string }>;
-};
 // these will only reload on page refresh (server only) or lang change
 const dictionaries: GetDictionary = {
   en: (namespace) =>
-    import(`./locales/en/${namespace}.json`).then((module) => {
+    import(`../locales/en/${namespace}.json`).then((module) => {
       return module.default;
     }),
   fr: (namespace) =>
-    import(`./locales/fr/${namespace}.json`).then((module) => {
+    import(`../locales/fr/${namespace}.json`).then((module) => {
       return module.default;
     }),
 };
 
-type GetTranslation = (
-  language: ValidLanguage,
-  namespace?: ValidNamespace
-) => Promise<
-  (k: string, params?: { [key: string]: string | number }) => string
->;
-
 export const getTranslation: GetTranslation = async (
   language = defaultLanguage,
-  namespace = defaultNS
+  namespace = defaultNamespace
 ) => {
   const dictionary = await dictionaries[language](namespace);
   return (key: string, params?: { [key: string]: string | number }) => {
