+++ apps/ui/src/i18n/locales/fr/common.json
