+++ apps/ui/src/i18n/locales/fr/second-page.json
